refactor: simplify is_optimal method by removing debug_print parameter

+++ cpu_rt/src/ray_tracer/hittable/oct_tree/operations/combine.rs
@@ -7,7 +7,7 @@ impl<T: Leafable> OctTree<T> {
         let offset = [offset.x, offset.y, offset.z];
         let other_size = offset
             .iter()
-            .map(|s| s.unsigned_abs() as u32 + other.size)
+            .map(|s| s.unsigned_abs() + other.size)
             .max()
             .unwrap();
